Conflicts:
	src/doc/en/documentation/api/api_v2/index/create%7Ccreate_an_index%7C1.md
	src/doc/en/documentation/api/api_v2/index/delete%7Cdelete_an_index%7C4.md
	src/doc/en/documentation/api/api_v2/index/exists%7Ccheck_if_an_index_exists%7C2.md
	src/doc/en/documentation/api/api_v2/index/list%7Clist_of_indexes%7C3.md

+++ src/main/java/com/jaeksoft/searchlib/util/FunctionTimer.java
@@ -36,7 +36,7 @@ public class FunctionTimer {
 	};
 
 	private static double globalTotalTime = 0;
-	private static long fullTotalTime = 0;
+	private static double fullTotalTime = 0;
 
 	public final static Mode MODE = Mode.TIME_ONLY;
 
@@ -149,7 +149,7 @@ public class FunctionTimer {
 		case FULL:
 			synchronized (ExecutionInfos) {
 				System.out.println("EXECUTIONS INFO DUMP");
-				long totalTime = 0;
+				float totalTime = 0;
 				for (ExecutionInfo executionInfo : ExecutionInfos.values()) {
 					System.out.println(executionInfo);
 					totalTime += executionInfo.totalTime;
@@ -164,8 +164,8 @@ public class FunctionTimer {
 		case TIME_ONLY:
 			globalTotalTime += (double) fullTotalTime / 1000;
 			System.out.println(StringUtils.fastConcat("Total time: ",
-					Long.toString(fullTotalTime / 1000))
-					+ " / " + Long.toString((long) globalTotalTime));
+					Double.toString(fullTotalTime / 1000))
+					+ " / " + Double.toString((long) globalTotalTime));
 			break;
 		}
 		if (reset)
